Move /api/v1/session.go to /types/session.go
