+++ api/v1/proxy/check/check.go
@@ -20,7 +20,7 @@ type CheckResponse struct {
 type Settings struct {
 	Client        api.Client
 	GetPath       types.GetVar
-	GetSession    api.GetSession
+	GetSession    types.GetSession
 	OnModifyInput shared.OnModifyInput
 }
