+++ api/v1/proxy/shared/defaults.go
@@ -1,6 +1,8 @@
 package shared
 
-import api "github.com/styrainc/styra-run-sdk-go/api/v1"
+import (
+	api "github.com/styrainc/styra-run-sdk-go/types"
+)
 
 func DefaultOnModifyInput() OnModifyInput {
 	return func(session *api.Session, path string, input interface{}) interface{} {
