+++ api/v1/proxy/shared/types.go
@@ -1,5 +1,7 @@
 package shared
 
-import api "github.com/styrainc/styra-run-sdk-go/api/v1"
+import (
+	api "github.com/styrainc/styra-run-sdk-go/types"
+)
 
 type OnModifyInput func(session *api.Session, path string, input interface{}) interface{}
