+++ rbac/v1/proxy/delete_user_binding/delete_user_binding.go
@@ -3,7 +3,6 @@ package delete_user_binding
 import (
 	"net/http"
 
-	api "github.com/styrainc/styra-run-sdk-go/api/v1"
 	"github.com/styrainc/styra-run-sdk-go/internal/utils"
 	rbac "github.com/styrainc/styra-run-sdk-go/rbac/v1"
 	"github.com/styrainc/styra-run-sdk-go/rbac/v1/proxy/shared"
@@ -14,7 +13,7 @@ type DeleteUserBindingResponse struct{}
 
 type Settings struct {
 	Rbac           rbac.Rbac
-	GetSession     api.GetSession
+	GetSession     types.GetSession
 	GetId          types.GetVar
 	OnBeforeAccess shared.OnBeforeAccess
 }
