+++ rbac/v1/proxy/get_roles/get_roles.go
@@ -3,7 +3,6 @@ package get_roles
 import (
 	"net/http"
 
-	api "github.com/styrainc/styra-run-sdk-go/api/v1"
 	"github.com/styrainc/styra-run-sdk-go/internal/utils"
 	rbac "github.com/styrainc/styra-run-sdk-go/rbac/v1"
 	"github.com/styrainc/styra-run-sdk-go/types"
@@ -15,7 +14,7 @@ type GetRolesResponse struct {
 
 type Settings struct {
 	Rbac       rbac.Rbac
-	GetSession api.GetSession
+	GetSession types.GetSession
 }
 
 func New(settings *Settings) *types.Route {
