+++ rbac/v1/proxy/list_user_bindings/list_user_bindings.go
@@ -3,7 +3,6 @@ package list_user_bindings
 import (
 	"net/http"
 
-	api "github.com/styrainc/styra-run-sdk-go/api/v1"
 	"github.com/styrainc/styra-run-sdk-go/internal/utils"
 	rbac "github.com/styrainc/styra-run-sdk-go/rbac/v1"
 	"github.com/styrainc/styra-run-sdk-go/rbac/v1/proxy/shared"
@@ -17,7 +16,7 @@ type ListUserBindingsResponse struct {
 
 type Settings struct {
 	Rbac       rbac.Rbac
-	GetSession api.GetSession
+	GetSession types.GetSession
 	GetUsers   shared.GetUsers
 }
