+++ rbac/v1/proxy/list_user_bindings_all/list_user_bindings_all.go
@@ -3,7 +3,6 @@ package list_user_bindings_all
 import (
 	"net/http"
 
-	api "github.com/styrainc/styra-run-sdk-go/api/v1"
 	"github.com/styrainc/styra-run-sdk-go/internal/utils"
 	rbac "github.com/styrainc/styra-run-sdk-go/rbac/v1"
 	"github.com/styrainc/styra-run-sdk-go/types"
@@ -16,7 +15,7 @@ type ListUserBindingsAllResponse struct {
 
 type Settings struct {
 	Rbac       rbac.Rbac
-	GetSession api.GetSession
+	GetSession types.GetSession
 }
 
 func New(settings *Settings) *types.Route {
