+++ rbac/v1/proxy/put_user_binding/put_user_binding.go
@@ -3,7 +3,6 @@ package put_user_binding
 import (
 	"net/http"
 
-	api "github.com/styrainc/styra-run-sdk-go/api/v1"
 	"github.com/styrainc/styra-run-sdk-go/internal/utils"
 	rbac "github.com/styrainc/styra-run-sdk-go/rbac/v1"
 	"github.com/styrainc/styra-run-sdk-go/rbac/v1/proxy/shared"
@@ -16,7 +15,7 @@ type PutUserBindingResponse struct{}
 
 type Settings struct {
 	Rbac           rbac.Rbac
-	GetSession     api.GetSession
+	GetSession     types.GetSession
 	GetId          types.GetVar
 	OnBeforeAccess shared.OnBeforeAccess
 }
